# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mock Sensor Hub API
  description: Demo REST API for the diploma REST client (sensors + device config).
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development (port may auto-increment if busy)
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  uptime:
                    type: integer
                    description: Process uptime in seconds
  /sensors:
    get:
      summary: List sensors (paginated, filterable)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - in: query
          name: q
          schema:
            type: string
          description: Case-insensitive filter on sensor id or unit
      responses:
        '200':
          description: Paged sensor list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorListResponse'
    post:
      summary: Create or update a sensor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorUpsert'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /config:
    get:
      summary: Get hub configuration
      responses:
        '200':
          description: Config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubConfig'
    put:
      summary: Update hub configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HubConfigUpdate'
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubConfig'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Sensor:
      type: object
      required: [id, value]
      properties:
        id:
          type: string
        value:
          type: number
        unit:
          type: string
    SensorUpsert:
      type: object
      required: [id]
      properties:
        id:
          type: string
          minLength: 1
        value:
          type: number
        unit:
          type: string
          nullable: true
    SensorListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Sensor'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
    HubConfig:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
    HubConfigUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
    Error:
      type: object
      properties:
        error:
          type: string